[chore] getting started with stripe connect
Stripe Connect is built to support the marketplace functionalities we
are trying to build. The code has been copied from the tutorial app
stripe built to demonstrate how to build a marketplace.

We are sticking with nodejs to be lean and fast. I have the most
experience building applications with nodejs. In the long run we want to
open source this technology or repo configuration therefore we want a
platform that a large amount of people can leverage. The support from
npm would make this higly customizable as well.

https://github.com/stripe/stripe-connect-rocketrides