[chore] create a server service
We want to split up the codebase into services to make it more
managable. Here we house all the backend logic. In time this will
grow and we will split it out into more services. However there must
be great care with how the data is handled when we do this in the
future.