[feat] create the client service
The client users will interact with on the platform is a product of the
Create-React-App project. Using React with Webpack will allow us to port
over most of the Gatsby Manor styles to babylon.

There will need to be an upgrade to Webpack v4 when their integration PR
is finally merged.